Moved dummies module to better location.
